backend/requirements.txt
mcp
httpx
finnhub-python
anthropic
python-dotenv
pytest
beautifulsoup4
readability-lxml
lxml
google-genai
pydantic
langchain-core
langchain-google-genai
langchain-anthropic
langchain-groq
langgraph
fastapi
uvicorn[standard]
# Self-contained dependencies for the Risk Pipeline
fastapi
uvicorn[standard]
pydantic
pandas
numpy
scikit-learn
torch
transformers
lightgbm
joblib
pyyaml
requests
httpx
groq
google-generativeai
python-dotenv
gunicorn
tavily-python
lime